+++ templates/database/migrations/2024_03_07_175856_create_terms_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    /**
+     * Run the migrations.
+     */
+    public function up(): void
+    {
+        Schema::create('terms', function (Blueprint $table) {
+            $table->id();
+            $table->timestamps();
+            $table->string('title');
+            $table->timestamp('term');
+            /// in minutes:
+            $table->integer('duration');
+            $table->text('description')->nullable;
+            $table->integer('visibility_scope');
+            $table->foreignId('owner_id')->references('id')->on('users')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
+    {
+        Schema::dropIfExists('terms');
+    }
+};